Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  f73209dc946d306357ed6dff90c0a9ad1de0c263
QA: make wikitext editor test pass in Chrome

Sometimes it takes a moment for an element to disappear, and
in this case Chrome was checking so quickly that the
"should_not be_visible" line was failing.

This change is actually a pattern you'll find throughout
the browser tests in the repos.

Change-Id: Iecfea2ec303f02f4a145a3320566641f2bde7064