Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  9246b29a33cd6a33af426a70ad5c018af4bf487c
Hygiene: Prevent warning when Echo is not installed

Previously prepareUserButton() would try to get the title of the
Notifications special page regardless of whether Echo was
installed. Now it makes sure that Echo is installed first. Rather
that splitting the default into 2 else statements, I just moved it
to the beginning.

Change-Id: Ia097a92fd88b740597faf736800ea812a920a14c