Updated mediawiki/extensions Project: mediawiki/extensions/Flow  bcb08f91c9c0e10e590b02bf9ffe976442564737
Standardize URL generation

Which parameter to use for which action was spread throughout
the application. This consolidates that into a single location.

This intentionally leaves behind PostActionMenu's usage of
generateUrl, the entire PostActionMenu will be deprecated with
the new frontend and is not worth figuring out how to convert over.

I think this patch is still an excellent step forward and should
be merged, even with the above caveats.

Change-Id: I00c8390b5631557f058f1c0fa3cd5c364d502872