Updated mediawiki/extensions Project: mediawiki/extensions/Flow  edbe89f790e096130b1bd394b24b32be1b4700dd
Remove some of the magic in formatTimestamp

Unfortunately, there seems to be little consistency in how all data is formatted
in all those places (and with Flow having different functionality, we'll likely
be adding more inconsistency)
I've removed the magic of "topic" being removed from $data['links'], and the
magic of auto-selecting the last available link. Instead, a default selection
of 'topic' and 'board' link will be picked.

The "topic" being removed from links didn't make sense anymore because I've
made changes to what links need to be displayed in the patch this one depends on.

Meanwhile, I've added a 'board' link - if posts & topic link to the topic,
header changes should link to the board (this will likely change in the next
patch, where timestamp will link to that specific revision)

Change-Id: Ib50138eb46cca479cf06b7e96eaef476c39d234b