Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  88752b940c1fe25d609cd017fc371d4e101c27ea
Reduce loading spinner toggle to a single 'loading' CSS class name

Originally I wanted to make the spinner a pure CSS solution but
could not do that for now because it relies on the
.mw-small-spinner class. We can address this in a later patch if
we want, e.g. by using our own spinner GIF.

Main advantage of this change is that it now supports multiple
Entities on one page.

Change-Id: I618977a80406267f309037914dea2afad10f8f47