Updated mediawiki/extensions Project: mediawiki/extensions/MediaWikiChat  6dfa078feb174dd9a238d09772b45085b4bfc562
Changing audio/ directory to assets/

What this does is allows for a directory for assets in general,
not just for audio. In the assets directory, there is a new
chatmod.png file. This is the new default chat-mod-image by default
instead of having chat-mod-image being an external resource from
images.brickimedia.org. It is no longer Brickimedia-specific and is
a generic star icon instead of a golden LEGO minifigure.

Not changed in this commit is all the i18n files. The en.json file is
changed to reflect the new icon URL, but I have not changed the
other i18n files as I assume they will be adjusted on translatewiki.net

Bump version no.

Change-Id: Ibcfd9ae334dcc5a0454ef00a1acdf40d6cd42fa2