Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0c3182cf08c9a4f819c4c523f7b777e73a05204b
Use SimpleSiteLink (for now) in UpdateRepo

We do need to eliminate usages of SimpleSiteLink, but
this provides workaround, for now, to avoid fatal error
that sometimes occurs when running phpunit tests.

"Fatal error: Cannot use Wikibase\DataModel\SiteLink as SiteLink
because the name is already in use"

We can then sort out usage of SimpleSiteLink in Wikibase.

Bug: 63407
Change-Id: Id80823f64d4d49b1b292061cf9c367409b2fc5c5