Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  932abe82a2a626a2e597c548e831b4ff45901ed2
Do not fraud check authorize with full amount

Because we have to have separate authorization with fraud
check and authorize with payment steps; we should not attempt
to authorize for the full amount.

Basically a donor will see 3 transactions in their card statement
so we should make it as non scary as possible with small amounts.

Change-Id: I2e51f0e37f1f548d5a7aa19c49c0c8d7b8c6bd59