Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  137dc6dcb9c641dbed448067ceb90187830b533f
wfStripIllegalFilenameChars truncates title

wfStripIllegalFilenameChars doesn’t just replace /, \,
it removes any part of the string before it. some titles contain these characters
as part of the metadata value; e.g., when the value is a URL. we don’t want to
truncate those strings; instead we want to preserve the legal characters.

Bug: 62909
Change-Id: Id48250851d5f540156b03567452972e918ca3b3b