Updated mediawiki/extensions Project: mediawiki/extensions/SyntaxHighlight_GeSHi  b6d8b1a5b9a75a085d7d609db3524d99b99d345c
Use ResourceLoader modules for style loading

We want to be able to track what styles were added to be able to deliver
this information to MediaWiki's live preview functionality (in order
to solve bug 24134).

This required moving some code in SyntaxHighlight_GeSHi class around.

The old way still works and is used for MediaWiki 1.20 and lower.

Bug: 24134
Change-Id: Iafd91de8922be55688fedef4e43a8e7f54d4e1cc