Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  c9d3c2c58b178f34fc2a2236d449bd510be0c3fa
Limit number of fragments scored if possible

Since the best fragments are typically at the beginning of the article any
way we can relatively safely stop searching for matches after 50 fragments.
This should help with really crazy documents, say 10MB of "d d".  Without
this we'll scan out all the "d"s on a search for "d". With it, only the
first 50.

Only works for experimental highlighter because, well, only the experimental
highlighter has this feature.

Change-Id: I09b9718ee84fb4cf30178b9b2949f55513b3f448