Updated mediawiki/extensions Project: mediawiki/extensions/Flow  68259be6979994f56156641e96647ec341255a06
Remove period from end of revision comments

In the history pages the separator used is

    <span class="mw-changeslist-separator">. .</span>

When combined with a period at the end of each revision message
Flow entries display with `. . .` rather than core history
entries which only have the double `. .` as expected.  For
consistency sake remove the extra dot.

Change-Id: Ibd6b510ed2a761006c05deb060941ee09da644eb