Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  bcf2d7b829a801ca420f9090581f29fd4cca0dcb
AVS Forcing and Removal of Server Address Check

Adding the globalcollect / adyen address and ZIP staging functions
to always force the AVS check when those fields are empty.

(Do this by unifying all the stage_(zip|address) functions.)

Additionally; allow empty address fields on the server because we
do not currently have a method of knowing when those fields are
actually required! So; we will rely on the client side check and
fall back to AVS failing when the fields are empty and we need them.

Change-Id: I97e5069857410a762ecec8b7797240baeb88100b