Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  aed691b3ee6aa466fded59e1c87b104be82f6bb0
Use some analysis plugins if installed

This adds support for Hebrew, Polish, Japanese, and Chinese plugins.

DEPLOYMENT: This does nothing if the plugins aren't installed and even
after that the index will have to be rebuilt to use the plugin.  Once
it has been rebuilt, though, removing the plugin will break stuff.

Change-Id: Ic9a3e6b5168ce777df0dbd4a6839b2f2f11b4fa2