Updated mediawiki/extensions Project: mediawiki/extensions/MediaWikiChat  bfd7aa3230c28ca2a34ca8fe8ff03e04a57a058a
Fixing private message links so they open in new tabs

For some reason, links, both internal and external, in private messages don't
open in new tabs. This can be frustrating especially if when you click a link
and it loads in the same tab as Chat, causing you to lose the entire chat
conversation. This commit fixes this in the same way I fixed links in chat-topic.

Change-Id: Icda3ef7de5ed6063201b7e30e899324ec83ca167