Updated mediawiki/extensions Project: mediawiki/extensions/GettingStarted  22fbeedfc51fdd3ed780ed52bc880df8d07bfc19
Revert GettingStarted tokens logic

Parts of this were already reverted as an emergency in 1.24wmf1/2
because of #64573 and subsequently fixed in this lineage with 0477aa,
but it was discovered that this whole series is breaking the site in a
different way:

Our caching layer (Varnish) is configured to generally ignore
MediaWiki's Vary: Cookie directive. This allows it to keep one version
in its cache, rather than one per user, since Cookie values are
generally unique. The semi-obvious exception to this is when the Cookie
value contains either the word "Session" or the word "Token", as this
would make them "non anon" pages with a different, possibly private,
content (and Cache-Control would also avoid them being cached
altogether)

This changeset introduced a gettingStartedUserToken Cookie for all
anons, which made Varnish just keep different versions for*every (just
invalidated) page for *every* user in its cache. This would already be
bad in terms of hit ratio and backend load, but it got even more
spectacular by a combination with a Varnish "limitation" during which
Varnish was crashing when attempting to purge pages that had accumulated
thousands of variants (i.e. certain popular pages, such as Main_Page).

The issue is being worked around at the Varnish layer with
Ice1648d1c240af087992f31781887f561f55b288 and changing the Cookie name
here would also be enough to make this a more sane change, but since a)
the data for experiment are probably already ruined by now and b) it's
the second different outage this individual change brings in 24h, I'm
opting in a full revert of the change for now and until it gets
reintroduced with a more thorough review.

This reverts commits:
 - 533c7f3b5821f6637f349dd0fd8ecd6a0de660d0
 - 99286bb24e86185fa7a0759588ec09ee74b2cd8c
 - 0477aa472a6a49f943097d6431ea7c44ecb310ca

Change-Id: I978282c1a83620b8522399fea9008d0379dce6df