Updated mediawiki/extensions Project: mediawiki/extensions/Echo  c9cff5c6002daa35df7d45af72811384e3e134e1
Always show the overlay when badge clicked

The badge is always close to the top of the window (if it's visible), so
this is not necessary. Also, this behavior is definitely not desired
for the upcoming fixed header beta feature (the badge will always be
visible as a part of the fixed header).

Change-Id: I66e8a50b1139f7bcd005cfb3c3576578efe6a653