Updated mediawiki/extensions Project: mediawiki/extensions/CodeEditor  96afba23d6630808a44e3b58e8f9b40bfbcf9a45
IE8: Fix line endings in CodeEditor

IE8 uses innerText, which is a bit more sloppy with line ending
preservation than other browsers are. Instead of copying the value into
the DOM and have the Editor pick it up during init, just set it directly
on the EditSession. Should even be faster.

Bug: 64559
Change-Id: Iafbb07afb6d06ed94057f8ddf8b080fd136d7fb3