Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5ba867cfcdb4a2615b6e8bc43454fd0ea208dadd
Fix selector for linkitem init

The span either has wb-langlinks-edit if the page already is
connected to an item or wb-langlinks-add if it's not.

But the span never has wbc-editpage which is now being set on
the a within the span. If this anchor is being used, we will
have two nested links (only one visible though) with two pencils
next to them.

Bug: 65001
Change-Id: Iac189e7513fdb366aa418dac891f61b51a898354