Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  acb72f379a4838e9f8c5695cadca37ff32a50bdd
Rearange array of test cases in MwTimeIsoFormatter tests

This is a prerequisit for one of my following patches. The reason
for the change is that neither the original ISO time stamp nor the
formatted time can be used as a unique key. It must be possible to
have two slightly different input strings that result in the same
output (e.g. "+2014" and "+000000002014" should be formatted as
"2014") as well as an input that results in two different outputs
(if the precision is different).

This patch does not add, removes or changes tests.

Change-Id: Ib1057f988b97d05b10db01e95c24000fbc43c56e