Updated mediawiki/extensions Project: mediawiki/extensions/Popups  b656841f8f185ca575bacd7930697f49d721dd69
Don't underline hovercard text when user preference is 'Always'

The setting is called 'underline' and defaults to 'Skin or
browser default' (= 2). If set to 'Always' (= 1), which is what
I personally like and do, ;-) all text in hovercards is
underlined (and the underlines disappear if you hover the
hovercard, which is a bit tricky but possible).

This simple fix removes the dependency from the default style
(which only shows underlines on hover). I think this is the most
straightforward solution.

Change-Id: Ide54ffd1949a50184d8d2a680bf1a0d35e24d563