Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  fb91f3e6bb9aca09f23a003ab3dade98cd1ebc64
Stop prefix: filter from searching redirects

Even though it is named after regular prefix searches, we don't want it
to search redirects.  Mostly, we don't want it to to that because people
don't expect it to.  Additionally, it is _very_ hard to figure out what
is going on because the prefix filter doesn't get highlighted.

Bug:  65232
Change-Id: I8c155f189c068631828faa59e1a84610eeba68a5