Updated mediawiki/extensions Project: mediawiki/extensions/PdfHandler  261af50ba625483a2200077dd7c42e5c76cab36f
Make validation for page more strict on pdf to take only numbers

This change causes wiki syntax like
 [[File:Foo.pdf|thumb|Page 7 of document]]
to be interpreted as a caption instead, of saying select page 7
of the pdf. Previously it eventually ran intval( '7 of document' ),
so flipped to page 7.

Only possible downside I could see is this would cause things like
left-to-right marks and weird unicode spaces to no longer be ignored.
I don't think that's a big deal.

Change-Id: Ib98510a0473458fdc9cdecdb7f75676488b4c5c8