Updated mediawiki/extensions Project: mediawiki/extensions/PagedTiffHandler  168c47bfde3c6a11458e2beb273aa37be0a051d2
Make validation for page more strict on tiff to take only numbers

This change causes wiki syntax like
 [[File:Foo.tiff|thumb|Page 7 of document]]
to be interpreted as a caption instead, of saying select page 7
of the tiff. Previously it eventually ran intval( '7 of document' ),
so flipped to page 7.

Only possible downside I could see is this would cause things like
left-to-right marks and weird unicode spaces to no longer be ignored.
I don't think that's a big deal.

Change-Id: I4760df0a230affb3189ca9b5f3f9f4cc76292b97