Updated mediawiki/extensions Project: mediawiki/extensions/VectorBeta  81de26ec7bde4a098bd8f53dcd737e446fc3498c
Fix Opera 12 flyout collapsing to early

This is a possible fix for a problem in Opera 12 (and possibly other
browsers) caused by the active hover reagion limited to max-height: 2.2em.

This second attempt is a bit different from the first (avoids px margins
and uses line-height instead) and tested in Chrome, Firefox and Opera.
The main purpose of this patch ist to let you know it's fairly easy to
avoid this problem.

I'm open for discussions.

Feel free to take this patch and submit additional patch sets on top
of mine.

Change-Id: I9bac959211e4f3d9bfb80793b153e3be3fb8c5db