Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5bc3d5dd98ef4f0c995ebabfff12f909dc35bdbd
Removed AutocompleteInterface

The change set is meant to be part of an iterative refactoring.
AutocompleteInterface was exclusively used by SiteIdInterface and SitePageInterface. However,
most of its functions were overwritten in SiteIdInterface / special to the derived interfaces.
Ultimately, there is no reason to keep AutocompleteInterface since the "old UI" components
are to be removed anyway. By improving the separation of SitePageInterface and SiteIdInterface,
a new version of the suggester / siteselector widget can be applied more easily.
The change should have no impact on the actual UI behaviour.

Change-Id: Iab9d6b7d6110825a034b45a4b96d89f4686eacb5