Updated mediawiki/extensions Project: mediawiki/extensions/Lingo  f9c28cca9254545b10366f68e42670bb089ff7a3
Unstrip strip items of 'general' category before lingo-parsing

The unstripping will be done again by the MW parser when the hook used by Lingo returns, but it should not hurt to do this twice.
The only problem is with other hook handlers that might not expect strip items to be unstripped already.

Bug 55829

Change-Id: I422650da84c07122451e19bf3424ad14f55d07af