Updated mediawiki/extensions Project: mediawiki/extensions/MediaWikiChat  18d79906787bdce6c70efe607a5e9b50a8a5f109
Adding a border to the bottom of the chat-topic

This aesthetic change makes the chat topic and the chat table have a division
between each other so that they don't feel like one thing. Also in this change
are some CSS cleanup per https://www.mediawiki.org/wiki/CC/CSS though these
changes in no way change the functionality or appearance of the extension.

Change-Id: Ibc98b2436ed37c91f9199cfe4ce321a12ef6c76e