Updated mediawiki/extensions Project: mediawiki/extensions/Math  d93fdf914157e37709cf1c09f77a372b0eadedf6
Add separate database table for LaTeXML (step 1)

Currently the PNG and the LaTeXML rendering mode
use a common table in the database. If both rendering
modes are allowed in a wiki simultaneously this causes
problems, because the fields are continuously overwritten.

This change adds a new table for the LaTeXML rendering mode
to the database.

Bug: 65522
Change-Id: If5b887e908f22248789d478c34b03ea3c7da393a