Updated mediawiki/extensions Project: mediawiki/extensions/Math  972cee0f98c0808985144012245c6a11d723ce7d
Add separate database table for LaTeXML (step 2)

Currently the PNG and the LaTeXML rendering mode
use a common table in the database. If both rendering
modes are allowed in a wiki simultaneously this causes
problems, because the fields are continuously overwritten.

This change creates the new database table defined in step 1.

Bug: 65522
Change-Id: I5f125d36b83776ca6f75ff014dc9095d0a737a35