Updated mediawiki/extensions Project: mediawiki/extensions/SemanticForms  b6a888b49adcc163831ce208e74d4f693046cb4b
Add PHPUnit boostrap

This adds a bootstrap as well as a convenience test runnner to invoke the
MW test environment.

Just use `php mw-phpunit-runner.php [options]` to run the tests.

On a general note, tests should be as independent as possible, using
`MediaWikiTestCase` will create an additional layer of dependencies
therefore it is suggested to use `PHPUnit_Framework_TestCase` instead
and only if a DB connection is required to fall back on
`MediaWikiTestCase`.

Change-Id: I2575b77e83f3788f150285dccc811c898e86384e