Updated mediawiki/extensions Project: mediawiki/extensions/PagedTiffHandler  3ed2aced63b7559db0255c6ab39664d5bc704c18
Make tiff thumbnail in 2 steps.

Tiffs can be really big to transfer around all over the place, the
largest one at commons is 1 gb. So instead render first to ~2048px,
and then downsize from that.

Vips has a quirk where it is much more memory efficient if the
factor we are shrinking by is precisely an integer. In order to
take advantage of that, this doesn't use core's bucketting system.

This only uses a single intermidary step. In the future it could be
extended fairly easily, but I think it would be good to start here,
and evaluate if a more complex system would be useful after doing
just the single step.

I've tested this patch with the largest resolution tiff file
currently on commons ([[File:Zoomit2.tif]]). Locally it renders
that file within the resource limits that production has.

This change requires Id3a8b25a598 in core.

Bug: 52045
Change-Id: Ifdcb93ea9198f32f0684cf90d4214afe09051000