Updated mediawiki/extensions Project: mediawiki/extensions/Flow  5e48ac80798064e248a6fa8e2d2b80eb123e9b8c
Be more forgiving in maintenance script

I pulled DB data from ee-flow & set it up on my local machine.
I wanted to run all patches again and ran into this error.

Probably got there because some local data corruption, so I'm
not really going to figure this out.

Regardless, if we can't find the rev_type_id, we can just let
the script continue. We won't be in worse shape after it: the
type id can't be found and the revision is lost already...

Change-Id: Ie93d06cd17c7493f0155121dfffd7b19c66972a4