Updated mediawiki/extensions Project: mediawiki/extensions/Math  423d80ef358a4ceaf9f3ef66f6b0eae58358cf1f
Add separate database table for LaTeXML (step 3)

Currently the PNG and the LaTeXML rendering mode
use a common table in the database. If both rendering
modes are allowed in a wiki simultaneously this causes
problems, because the fields are continuously overwritten.

This change adds the method getMathTableName that will
allow renderers to access their specific table in the
database.

Change-Id: I8b11ffd0a39fadd7d3c8de1e94e74ff5f490430d