Updated mediawiki/extensions Project: mediawiki/extensions/Echo  e205992d0bba9c8b5d5882b16e57763cfd852342
Remove unused code and fix missing/broken documentation

General code cleanup as reported by the PHPStorm static code
analysis. I hope it's not a problem that I made a lot of very
different (but all very tiny) changes in a single patch. If you
want to merge this but you think it's better to split it into
several patches first, please tell me.

Change-Id: I2e2c4bb47f8d20e038d28e236e2ff813b30504af