Updated mediawiki/extensions Project: mediawiki/extensions/Math  410416101f80cd052f82a8bbdfc49d4dbac2b93b
Add separate database table for LaTeXML (step 4)

Currently the PNG and the LaTeXML rendering mode
use a common table in the database. If both rendering
modes are allowed in a wiki simultaneously this causes
problems, because the fields are continuously overwritten.

In this change the MathRenderer class fields are updated
and new setters and getters are introduced.

Bug: 65522
Change-Id: I09ba07c5ed6b16d65cffe4da238d138a6cafd795