Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  19202aa8ee832a698b4e14c74899a39b95f72c92
Unify imports of namespace-less classes

Optimizing imports in PHPStorm (Ctrl+Alt+O) removes the line
"use MediaWikiTestCase" for some reason. This may be a bug in
PHPStorm. But this doesn't matter. It seems so common that all the
tests we have use the syntax "extends \MediaWikiTestCase" to avoid
this problem. Except for these few tests. Which I made equal now.

Change-Id: I097cbe306156282768fcaa559f7d55628436a6ee