Updated mediawiki/extensions Project: mediawiki/extensions/GettingStarted  1290c8ee0e3ba6da0260cf73351749fae17ae047
Optionally depend on CentralAuth's redirect hook

Dependent on core hook PostLoginRedirect, which allows for modification
of redirect url.

Note:
Without CentralAuth running, the onPostLoginRedirect hook will be
registered and the user will not automatically be redirected as they are
with the C.A. redirect.  As per the wikis normal login flow they will
be presented with a return to link.  We could decide to auto redirect by
changing the type, but I think a discussion should be had prior.

Core dependency: I219353d9825918e22d4772611a38d72c4e674e10
Bug: 65619
Bug: 61758
Change-Id: I732de20f5555ef96d2f3b4a87df46e04e38e8493