Updated mediawiki/extensions Project: mediawiki/extensions/Scribunto  b82ed4aa7d9e313c39e87e064cfd1d8218167be6
Restrict cached results to their original frame

When caching results from frame:preprocess and frame:expandTemplate,
restrict the scope of the cache to the frame object that was used. This
allows the integrity of the empty-frame expansion cache to be maintained
while also allowing parent frame access. This change is the equivalent of
I621e9075 in core.

Change-Id: Iae4c00e7e19ba12cfdaac135be16c991d9d0cea1