Updated mediawiki/extensions Project: mediawiki/extensions/Mantle  343539f163058493eb5c34a219edf684837539a7
Remove dependency on ext.mantle.oo to Hogan

The OO library should be and is template agnostic.
This makes this clear and doesn't pull in 2 template libraries unless
you try to use 2 different types of templates

Fix the compile function so the 2nd argument has an effect
update View test so template agnostic.

Change-Id: Ic98a2d63c0af0a3d945fdc20f368f4231cc620b7