Updated mediawiki/extensions Project: mediawiki/extensions/TemplateSandbox  6ef3685983a3d80d0fa27806057a754074b9204f
Properly handle preview when creating a template

WikiPage::replaceSectionContent returns null if the page doesn't exist
and a section edit is being attempted. When the template doesn't exist
yet, we need to do the same as EditPage does and just use the textbox
content as-is, possibly with the section header added.

Bug: 66141
Change-Id: I35e0b643d9ae0e8f6b2cfb484786b57929c58d1e