Updated mediawiki/extensions Project: mediawiki/extensions/Math  644513651f71c2e38b27e7f9e1f3f1e77c9e7fa5
Add getHtmlOutput method

Currently the method render always returns
a html string that can be a string that
represents the correct result or a rendered
error message.
This change adds a mechanism that allows
to fetch the HTML output.
In a followup commit the rendering function
is changed to return a boolean value rather
than the rendering result.
This will simplify the error handling and caching.

Change-Id: I80760493e391911c41eb69d75a93c6a34db8852e