Updated mediawiki/extensions Project: mediawiki/extensions/SemanticForms  396bacb6bd5d555821c274760993b54d5d32dbf0
Combobox switched to Select2 autocompletion

Combobox now uses Select2 JS for autocompletion.
A new global variable $sfgMaxLocalAutocompleteValues is
introduced, which is settable via LocalSettings.php, it
is used to automatically switch to remote autocompletion,
if the number of values is more than $sfgMaxLocalAutocompleteValues.
Semantic property field of CreateTemplate is also switched to
Select2 combobox.

Bug: 61974
Change-Id: I069f06dfbb912843cc0f97f1b31a32c57f5c4a08