Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  7c3ff509d64c7a4e4b9034a005824d14f7ee3003
AntiSpoof for CentralAuth: Update extension credits url

This plugin doesn't ship with AntiSpoof, it ships with CentralAuth.

Since this is provided by the CentralAuth extension, the url
should point there. I was trying to locate the source code of
"AntiSpoof for CentralAuth" when I found this thing listed on
Special:Version. Ended up on AntiSpoof's documentation page
which links to a git repo containing only AntiSpoof.

Change-Id: If1c0dbc7ad300873c93d74fc5f2369db5da00955