Updated mediawiki/extensions Project: mediawiki/extensions/BookManagerv2  3b037122d6c0e896e8e434dc3e4a3fcc964befe6
Import chapter list from existing book in the same order as on the contents page

Create an api module for returning internal links in the same order as they appear on the page.

Bug: 66367
Change-Id: Icac6a7a40925e6aef70d4b963b6ba476504d3141

Ease migration of existing books

Functionality to import table of contents to json format from an extisting contents page of a book.

Bug: 53286
Change-Id: I392a3c4bebbf8325d33db3b5d62d5d415d47da4f