Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  fbab814bd850f723246a84f6785ef9155b8891e8
A guided tour for publishing a translated page

The general structure and the i18n should be more or less done,
but the flow and the cookies need work.

This patch relies on the following core change:
I2765248f61ecd3089a9f4e06571a378e39ec1db3

It introduces two cookies:

1. A general GuidedTour cookie, named "(prefix)-mw-tour",
   with the usual GuidedTour info: tour name, start time,
   first step, etc.
2. A cookie specific to this extension, named
   "(prefix)-cx-published", which includes the username
   and the name of the translated page.

Both cookies are created when the user publishes
the first draft of a translated page, and both are deleted
when the user completes the page publishing by
moving it from the user space to the article space.

Change-Id: I15759019aea56a952eeb03aa25b83b9fb35e9c30