Updated mediawiki/extensions Project: mediawiki/extensions/ConfirmEdit  1e5bd7dc3c1bed29fea7a9209e9e30009d7e37e1
Add more words to blacklist

In adition to profanities, remove silly or questionable words
thus reducing the possibility of captchas like "killkids" or
"allahlies". I understand that this list is still nowhere near
complete, but at least this is a step in the right direction.

For the reference, the current wordlist used for captcha generation
is http://noc.wikimedia.org/~maxsem/words.txt
(words with anything but lowercase letters also get discarded)

Change-Id: Ic1cb0aa688811886e013bc819866abd745d02f68