Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  193f6a4f0636a991658ed87001a33ce6b4982bc1
jquery.payment: don't express dependency on jQuery

jQuery is a dependency for ResourceLoader itself, so you can be sure it has
always already been loaded. If you try to express a dependency on it
explicitly, a test in core will fail:

26) ResourcesTest::testDependencies
Module 'jquery.payment' must not depend on 'jquery'
Failed asserting that an array does not contain 'jquery'.
in tests/phpunit/structure/ResourcesTest.php:50

Change-Id: I550d89e94550e45e72200152d9316f8b01541b2e