Updated mediawiki/extensions Project: mediawiki/extensions/Flow  e64dc171d18c35a305c50de2cedf12af54022b56
i18n: Rephrase the message flow-topic-participants-second-try

The current message uses {{PLURAL}} in an unusual way -
it concatenates a comma conditionally.

It's better to repeat a few words, even in English, because it
makes it clearer to the translators.

This commit also rephrases the plural parts a bit to make it
look less like comma splice.

Change-Id: Ie2f53c08bc1f23cab00b2b6cd28c53a79106fdd6