Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  939213dd2934c351649e046b99a42d46cb64f4ba
Enable or disable the publish button based on progress of translation

Checking presence of text was not reliable. The event on which this check
was done also not reliable.

For example, translating first paragraph and then selecting that
and deleting using delete key made the publish button active.

In this commit, we use the percentage of completion as the criteria
for enabling and disabling the button. If progress is at least 1%,
we enable the publish button.

Change-Id: I1acc9a19e89c6cecd25a918391bdf30fbf4ba82c