Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  bbd10a9e8c57307552c7ba93c08447511fc2f366
Avoid multiple "add translation" indicators at a time

By default we want to suggest the user to add a translation to the first paragraph,
but if the user hovers a different one, we don't want to show the indicator in both.
This patchset makes it possible to suggest the initial paragraph
only if the user is not hovering a different one.

Change-Id: I65587c0e8c7a28878c69bb235983c5c5f152ba85