Updated mediawiki/extensions Project: mediawiki/extensions/GraphViz  d9a02b0f4aea0af88554d908df50ba3c9a372d3d
Backward compatibility fix for MW version <1.21.

The problem is that the GraphViz extension 1.0+ uses two hooks and
one function that were introduced in MW 1.21 (PageContentSave,
PageContentSaveComplete and WikiPage::doEdit respectively).

The fix is for the extension to use the older hooks and function
for MW versions <1.21 (ArticleSave, ArticleSaveComplete and
WikiPage::doEditContent respectively).

Bug: 67587
Change-Id: Ie353c40ff784e49cb7740059c2e653443dfa0cce