Updated mediawiki/extensions Project: mediawiki/extensions/Flow  a1a0b36099fcc3c481b2bca874c6fd7af189d405
Generic error handler for edit conflicts

This will allow re-submission of stuff that failed to submit because of an edit
conflict.
When an error is discovered after submitting <content>, and when the API returns
the most recent revision id (of the newest, conflicting, revision), this new ID
will be stored. Later, when resubmitting, a generic pre-handler will be called
when this ID is present, which will swap out the ID that's automatically read
from the (at this point "outdated") form target & substitute it with the one
we've been fed on our failed attempt to save the edit the first time.

Change-Id: I7de788492908cdbd089b361eb9f08cd1844ed794