Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  78b081daddc3b635878204b651c2cbedaa3c9e28
Years are absolute, not relative (ago/in)

Just discussed this with Lydia and Daniel.

It was always awkward that these messages sounded like they are
relative to today (the year 2014). This is not a problem with big
numbers (1 million BC and 1 million years ago mean almost the same)
but is an issue with small numbers (10000 years BC and 10000 years
ago are different, right?).

This patch makes all messages absolute.

Change-Id: I45faeb54b937b5fca2fa050d8c692b701d957cd7