Updated mediawiki/extensions Project: mediawiki/extensions/Petition  ed3ed7c752fe5a3d10c5f06261bbf2191912b260
Add logging for signatures, and integration with CheckUser

Adds a private log (only visible to users with view-petition-data right)
that shows which users/IP addresses signed a petition. Intended to prevent
abuse, and allow blocking them if necessary.

If the CheckUser extension is installed, signatures will now be logged
to it also.

Bug: 65850

Change-Id: I81f17df9de01f97e141a96bb906388da664f32b6