Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  86a91e6637a6659bc65b45d8d92849549317d8c3
Fix for TMH "Error missing media source" under HHVM

Media players inline showed only as "Error missing media source".
Appears to be due to use of current() to get the first element of
an array, assuming that the key pointer was reset.

In HHVM, magic reset on pass-by-value doesn't appear to happen.
Adding a manual reset() before the current() usage fixes it.

Bug: 67937
Change-Id: I606048cca847ec4927e2a59c44d819dda067db3a